+++ src/post/Feed/PostFeedCard.scss
@@ -1,5 +1,3 @@
-@import "../../styles/colors";
-
 .PostFeedCard {
   border: 1px solid #E7E7EC;
   box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.05);
