+++ src/widgets/Flag.scss
@@ -1,0 +1,12 @@
+.Flag {
+  border-radius: 50%;
+  width: 24px;
+  height: 24px;
+  line-height: 1em;
+  vertical-align: middle;
+}
+
+.Flag--lg {
+  width: 48px;
+  height: 48px;
+}
