+++ src/widgets/Icon.scss
@@ -1,31 +1,31 @@
 @import "../styles/colors.scss";
 
-.icon {
+.Icon {
   vertical-align: middle;
   color: #AAA;
 }
 
-.icon-xs {
+.Icon--xs {
   font-size: 16px;
 }
 
-.icon-sm {
+.Icon--sm {
   font-size: 22px;
 }
 
-.icon-md {
+.Icon--md {
   font-size: 28px;
 }
 
-.icon-lg {
+.Icon--lg {
   font-size: 36px;
 }
 
-.icon-xl {
+.Icon--xl {
   font-size: 86px;
 }
 
-.icon-menu {
+.Icon--menu {
   height: 56px;
   line-height: 54px;
   margin: 0 14px;
@@ -36,28 +36,13 @@
 
 .main-panel {
   a:hover {
-    .icon {
+    .Icon {
       color: $primary-color;
     }
   }
-}
 
-.flag {
-  border-radius: 50%;
-  width: 24px;
-  height: 24px;
-  line-height: 1em;
-  vertical-align: middle;
-}
-
-.flag-lg {
-  width: 48px;
-  height: 48px;
-}
-
-.cell.cell-bottom {
   a.active {
-    i {
+    .Icon {
       color: $primary-color;
     }
   }
